serialize writes on each vnode to avoid divergence in writes
when a write to W vnodes is required it's sent to the first vnode in
the PrefList, this vnode sends a task to a child process it has
that spawns the fsm to run the write on the W vnodes, this way the
writes are queued in this task runner mailbox and run synchronously,
thus avoiding two writes at the same time that may arrive differently to
the W vnodes and cause divergence in order.

previosly the fsm that coordinated the writes was created from the
process that received the http request, this caused that potentially N
http requests running in parallel would fire writes to the same stream
and if the order of the writes interleaved it would result in the order
of writes to be different in the W vnodes that received the write
requests.

note: in the future we should use something like pobox[1] to drop write
requests if the writer process becomes to busy.

[1] https://github.com/ferd/pobox