only shutdown on channel terminate when listen is active (sse and ws), read below
this is my theory from what I've observed in the logs and adding more logs:

if doing comet on one channel and we are the only subscriber and using
keep alive connections the bullet process is the same across xhr
requests, when a subscription happens and we get a message we
unsubscribe in terminate, then since the smc channel has only us as
subscribers it terminates, there it sends a message to the subscribers
(just us) signaling that it terminated, but the current connection
already closed, a new one happens and has the same pid as the last one,
so it starts and receives a message saying that the channel closed and
then it used to decide to terminate, but it was an old message for the
old listener, not for us. Now we ignore terminates since we will
terminate after handling one message or timing out anyway on xhr.